CORE-16177 Addressing PR comments
